fixed a refactor bug
